Update XMTP SDK and refactor group chat management logic
