+++ backend/package.json
@@ -13,7 +13,7 @@
     "start": "tsx src/index.ts"
   },
   "dependencies": {
-    "@xmtp/node-sdk": "^1.2.0",
+    "@xmtp/node-sdk": "^2.0.6",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.21.2",
